Remove ghost number from data format
The goal is to remove the need to modify a recorded ghost between storage
and replay to adjust the ghost number.

Instead, each ghost will start with the 4 fields for private use
and version number, allowing each ghost to be recorded in a different
version, in line with requirements.

For playback, the list of 5 ghosts will be sent as an array of ghost records.

I will further reformat the text and add two examples before tagging
the version v0.2.0.